Move registration and login methods to separate controller
