+++ app/Http/Controllers/WebshopController.php
@@ -54,117 +54,6 @@ class WebshopController extends Controller
     }
 
     /**
-     * Show the registration page
-     *
-     * @return mixed
-     */
-    public function register()
-    {
-        if (Auth::check())
-            return redirect('/account');
-
-        $data = [];
-
-        if (Session::has('registrationData'))
-            $data = Session::get('registrationData');
-
-        return view('webshop.register', $data);
-    }
-
-    /**
-     * Verify the registration page
-     *
-     * @return mixed
-     */
-    public function register_check()
-    {
-        $validator = Validator::make(Input::all(), [
-            'corContactName' => 'required',
-            'corName' => 'required',
-            'corAddress' => 'required',
-            'corPostcode' => 'required',
-            'corCity' => 'required',
-            'corPhone' => 'required',
-            'corEmail' => 'required|email',
-            'corSite' => 'url',
-
-            'delAddress' => 'required',
-            'delPostcode' => 'required',
-            'delCity' => 'required',
-            'delPhone' => 'required',
-
-            'betIBAN' => 'required',
-            'betKvK' => 'required',
-            'betBTW' => 'required',
-
-            'digFactuur' => 'email',
-        ]);
-
-        if (!$validator->fails()) {
-            $data['corContactName'] = Input::get('corContactName');
-            $data['corName'] = Input::get('corName');
-            $data['corAddress'] = Input::get('corAddress');
-            $data['corPostcode'] = Input::get('corPostcode');
-            $data['corCity'] = Input::get('corCity');
-            $data['corContactPhone'] = Input::get('corContactPhone');
-            $data['corPhone'] = Input::get('corPhone');
-            $data['corFax'] = (Input::get('corFax') !== false ? Input::get('corFax') : "");
-            $data['corEmail'] = Input::get('corEmail');
-            $data['corSite'] = (Input::get('corSite') !== false ? Input::get('corSite') : "");
-
-            $data['corIsDel'] = Input::get('corIsDel');
-
-            $data['delAddress'] = ($data['corIsDel'] ? $data['corAddress'] : Input::get('delAddress'));
-            $data['delPostcode'] = ($data['corIsDel'] ? $data['corPostcode'] : Input::get('delPostcode'));
-            $data['delCity'] = ($data['corIsDel'] ? $data['corCity'] : Input::get('delCity'));
-            $data['delPhone'] = ($data['corIsDel'] ? $data['corPhone'] : Input::get('delPhone'));
-            $data['delFax'] = ($data['corIsDel'] ? $data['corFax'] : (Input::get('delFax') !== false ? Input::get('delFax') : ""));
-
-            $data['betIBAN'] = Input::get('betIBAN');
-            $data['betKvK'] = Input::get('betKvK');
-            $data['betBTW'] = Input::get('betBTW');
-
-            $data['digFactuur'] = Input::get('digFactuur');
-            $data['digOrder'] = Input::get('digOrder');
-            $data['digArtikel'] = Input::get('digArtikel');
-
-            Session::flash('registrationData', $data);
-
-            DB::table('registrations')->insert([
-                'company' => $data['corName'],
-                'formdata' => json_encode($data),
-                'created_at' => Carbon::now()
-            ]);
-
-            \Mail::send('email.registration', $data, function ($message) {
-                $message->from('verkoop@wiringa.nl', 'Wiringa Webshop');
-
-                $message->to('verkoop@wiringa.nl');
-
-                $message->subject('Webshop registratie');
-            });
-
-            return redirect('/register/sent');
-        } else
-            return redirect()->back()
-                ->withErrors($validator->errors())
-                ->withInput(Input::all());
-    }
-
-    /**
-     * Show the registration success page
-     *
-     * @return mixed
-     */
-    public function registerSent()
-    {
-        if (!Session::has('registrationData'))
-            return redirect('/register');
-
-        return view('webshop.registerSent');
-    }
-
-    /**
      * The product page
      * Will throw 404 error when no product matches the product id
      *
@@ -199,61 +88,6 @@ class WebshopController extends Controller
                 'prevPage' => $prevPage
             ]
         );
-    }
-
-    /**
-     * The user will be redirected to the previous page with
-     * a message indicating whether the login was successful or not
-     *
-     * @return $this|\Illuminate\Http\RedirectResponse|Redirect
-     */
-    public function login()
-    {
-        // Check if the user is already logged in
-        if (Auth::check())
-            return redirect('account');
-
-        // Is all the data entered
-        if (Input::has('username') && Input::has('password')) {
-            // Try to log the user in
-            if (Auth::attempt(['login' => Input::get('username'), 'password' => Input::get('password'), 'active' => 1], (Input::get('remember_me') === "on" ? true : false))) {
-                if (Auth::user()->cart) {
-                    foreach (unserialize(Auth::user()->cart) as $item) {
-                        // Check if the id is set to prevent problems when reloading the cart
-                        if (isset($item['id'])) {
-                            // Restore the user's cart
-                            Cart::add($item);
-                        }
-                    }
-
-                }
-
-                return redirect()
-                    ->back()
-                    ->with('status', 'U bent nu ingelogd');
-            }
-        }
-
-        // The input field(s) is/are empty, go back to the previous page with an error message
-        return redirect()->back()
-            ->withErrors('Gebruikersnaam en/of wachtwoord onjuist')
-            ->withInput(Input::except('password'));
-    }
-
-    /**
-     * The user will be redirected to the main page if the logout was successful
-     *
-     * @return \Illuminate\Http\RedirectResponse
-     */
-    public function logout()
-    {
-        if (Auth::check()) {
-            Cart::destroy();
-            Auth::logout();
-
-            return redirect()->intended('/')->with('status', 'U bent nu uitgelogd');
-        } else
-            return redirect('/')->withErrors('Geen gebruiker ingelogd');
     }
 
     /**
